Display real type information in the State view instead of always "Object" (#776)
* Display real type information in the State view instead of always "Object"

* Remove MD5 reference (was used for path hashing)
Object metadata cannot be based on paths, has to be based on reference, because the same object can appear at multiple paths
  - Luckily our serializer retains object references across serializations, so this can work for us

* Separate object metadata and component metadata into separate maps. We
cannot store component metadata keyed off values, because oftentimes
those values are scalars that are not unique keys that can be used in a
map. So now we store the component instance as the map key and then
retain a list of all properties and their metadata.